% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEasybreedeR.R
\name{run_dataprevieweR}
\alias{run_dataprevieweR}
\title{Launch dataprevieweR application}
\usage{
run_dataprevieweR()
}
\description{
Launches the dataprevieweR application for data preview and exploration.
}
\examples{
\dontrun{
  run_dataprevieweR()
}
}
