% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEasybreedeR.R
\name{run_easybreedeR}
\alias{run_easybreedeR}
\title{Launch the easybreedeR Studio Suite}
\usage{
run_easybreedeR()
}
\value{
A Shiny app object
}
\description{
Main entry point for the easybreedeR Shiny application suite.
This function launches the main studio application which provides
access to all sub-applications.
}
\examples{
\dontrun{
  run_easybreedeR()
}
}
