% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEasybreedeR.R
\name{run_easybreedeR_studio}
\alias{run_easybreedeR_studio}
\title{Launch easybreedeR Studio}
\usage{
run_easybreedeR_studio()
}
\description{
Launches the main easybreedeR Studio application (suite launcher).
}
\examples{
\dontrun{
  run_easybreedeR_studio()
}
}
