% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEasybreedeR.R
\name{run_easyblup}
\alias{run_easyblup}
\title{Launch easyblup application}
\usage{
run_easyblup(host = "0.0.0.0", port = NULL)
}
\arguments{
\item{host}{The IPv4 address that the application should listen on. Defaults to "0.0.0.0" to allow access from other devices on the network.}

\item{port}{The TCP port that the application should listen on. Defaults to NULL, which will use a random available port.}
}
\description{
Launches the easyblup application for BLUP/REML parameter generation
with AI assistant features.
}
\examples{
\dontrun{
  run_easyblup()
}
}
