% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEasybreedeR.R
\name{run_easybreedeR}
\alias{run_easybreedeR}
\title{Launch the easybreedeR Studio (suite launcher)}
\usage{
run_easybreedeR(host = "0.0.0.0", port = NULL)
}
\arguments{
\item{host}{The IPv4 address that the application should listen on. Defaults to "0.0.0.0" to allow access from other devices on the network.}

\item{port}{The TCP port that the application should listen on. Defaults to NULL, which will use a random available port.}
}
\value{
A Shiny app object (invisibly returned by \code{shiny::runApp}).
}
\description{
Main entry point for the easybreedeR Shiny application suite. Provides
access to all sub-applications (easyblup, pedivieweR, datavieweR, RCW).
}
\examples{
\dontrun{
  run_easybreedeR()            # launch suite
}
}
