% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEasybreedeR.R
\name{run_easybreedeR}
\alias{run_easybreedeR}
\title{Launch the easybreedeR Studio (suite launcher)}
\usage{
run_easybreedeR()
}
\value{
A Shiny app object (invisibly returned by \code{shiny::runApp}).
}
\description{
Main entry point for the easybreedeR Shiny application suite. Provides
access to all sub-applications (easyblup, pedivieweR, dataprevieweR, RCW).
}
\examples{
\dontrun{
  run_easybreedeR()            # launch suite
}
}
