% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEasybreedeR.R
\name{run_pedivieweR}
\alias{run_pedivieweR}
\title{Launch pedivieweR application}
\usage{
run_pedivieweR()
}
\description{
Launches the pedivieweR application for pedigree visualization
and quality control.
}
\examples{
\dontrun{
  run_pedivieweR()
}
}
